refactor: isolate update config snapshot imports
