+++ src/cli/update-cli/update-command-config.ts
@@ -1,9 +1,7 @@
 // Config snapshots and pre/post-update config restoration.
-import { existsSync } from "node:fs";
 import fs from "node:fs/promises";
 import { isRecord } from "@openclaw/normalization-core/record-coerce";
 import { theme } from "../../../packages/terminal-core/src/theme.js";
-import { createPreUpdateConfigSnapshot } from "../../config/backup-rotation.js";
 import {
   mutateConfigFileWithRetry,
   parseConfigJson5,
@@ -12,7 +10,7 @@ import {
 import { resolveConfigEnvVars } from "../../config/env-substitution.js";
 import { resolveConfigIncludes } from "../../config/includes.js";
 import { asResolvedSourceConfig, asRuntimeConfig } from "../../config/materialize.js";
-import { resolveConfigPath, resolveIncludeRoots } from "../../config/paths.js";
+import { resolveIncludeRoots } from "../../config/paths.js";
 import { parsePluginInstallRecordMap } from "../../config/plugin-install-record-map.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import type { PluginInstallRecord } from "../../config/types.plugins.js";
@@ -24,15 +22,6 @@ import { defaultRuntime } from "../../runtime.js";
 import { VERSION } from "../../version.js";
 
 const PRE_UPDATE_CONFIG_SNAPSHOT_MAX_AGE_MS = 6 * 60 * 60 * 1000;
-
-export async function createUpdateConfigSnapshot(
-  env: NodeJS.ProcessEnv = process.env,
-): Promise<void> {
-  await createPreUpdateConfigSnapshot({
-    configPath: resolveConfigPath(env),
-    fs: { writeFile: fs.writeFile, readFile: fs.readFile, existsSync },
-  });
-}
 
 export function normalizePluginInstallRecordMap(
   value: unknown,
