+++ src/cli/update-finalization-output.test-support.ts
@@ -80,8 +80,12 @@ const stubs = new Map<string, string>([
     "export const withPluginLifecycleLease = async (_options, run) => await run();",
   ],
   [
+    sourceUrl("./update-cli/update-command-config-snapshot.ts"),
+    "export const createUpdateConfigSnapshot = async () => {};",
+  ],
+  [
     sourceUrl("./update-cli/update-command-config.ts"),
-    "export const createUpdateConfigSnapshot = async () => {}; export const readPostCorePreUpdateSourceConfig = async () => undefined; export const persistRequestedUpdateChannel = async ({configSnapshot}) => configSnapshot; export const persistValidatedDowngradeConfig = async () => {}; export const restoreDroppedPreUpdateChannels = snapshot => ({snapshot, changed: false});",
+    "export const readPostCorePreUpdateSourceConfig = async () => undefined; export const persistRequestedUpdateChannel = async ({configSnapshot}) => configSnapshot; export const persistValidatedDowngradeConfig = async () => {}; export const restoreDroppedPreUpdateChannels = snapshot => ({snapshot, changed: false});",
   ],
   [
     sourceUrl("./update-cli/update-command-plugins.ts"),
